BoardConfig: Set default composition type
If we are running on 7x27A, debug.composition.7x27A.type will be set
for composition.
If we are running on 7x25A, debug.composition.7x25A.type will be set
for composition.
Set default composition type for 7x25A as mdp.
Set default composition type for 7x27A as mdp.

Change-Id: Ic2c5d383dd3987db32b20208d2353f9e127a76bb